fix(updater): widen statvfs result to uint64_t (32-bit overflow on pi32)
get_available_space() computed f_bavail * f_frsize in size_t, which is
32-bit on every armhf/MIPS32 target we ship (pi32, ad5m, ad5x, k1, k2,
sonicpad-armhf, snapmaker-u1, cc1). For any filesystem larger than ~4
GiB the product wraps mod 2^32.

Bundle D6LPLAYP made it concrete: a Pi 4 running 32-bit Raspberry Pi OS
on a 60 GiB rootfs with ~46 GB actually free reported '178.3 MB free'
across every candidate path, blocking the in-app update. The previous
fix (a183d5847) lowered the threshold floor from 200 to 120 MB, which
masked the symptom for users whose wrapped value happened to land above
120 MB — but the root cause was arithmetic, not headroom. This device
had 46 GB free, not 178 MB; 11,580,559 blocks * 4096 frsize wraps to
~181 MiB.

Widen get_available_space() to uint64_t with explicit casts before the
multiply, and propagate uint64_t through DownloadPathDiag and the
public get_download_path / required_download_space_bytes signatures.

Test: new [update_checker] case locks the multiplication contract using
the exact block counts from the bundle.

+++ include/system/update_checker.h
@@ -20,6 +20,7 @@
 
 #include <atomic>
 #include <chrono>
+#include <cstdint>
 #include <functional>
 #include <mutex>
 #include <optional>
@@ -192,21 +193,22 @@ class UpdateChecker {
     /// candidate met the free-space threshold, so callers can build a
     /// useful error message.
     struct DownloadPathDiag {
-        std::string best_dir;        // best candidate found (regardless of threshold)
-        size_t best_free_bytes = 0;  // its free space
-        size_t threshold_bytes = 0;  // threshold required
+        std::string best_dir;          // best candidate found (regardless of threshold)
+        uint64_t best_free_bytes = 0;  // its free space (uint64_t — on 32-bit
+                                       // platforms a 46 GiB rootfs wraps size_t)
+        uint64_t threshold_bytes = 0;  // threshold required
     };
     /// @param diag Optional diagnostic out-param (for error reporting)
-    /// @param threshold_bytes Required free bytes; 0 → use a 200 MB default
+    /// @param threshold_bytes Required free bytes; 0 → use a 120 MB default
     ///        for callers without a known download size.
     std::string get_download_path(DownloadPathDiag* diag = nullptr,
-                                  size_t threshold_bytes = 0) const;
+                                  uint64_t threshold_bytes = 0) const;
 
     /// Compute the disk-space threshold for an in-app download.
     /// Returns 1.2x download_bytes + a small buffer when known, else a fixed
     /// default sized for current release archives. Always honors a 50 MB
     /// safety floor.
-    static size_t required_download_space_bytes(size_t download_bytes);
+    static uint64_t required_download_space_bytes(uint64_t download_bytes);
 
     std::string get_platform_asset_name() const;
